+++ tests/test_util.py
@@ -299,14 +299,22 @@ class TestSaveConfig(object):
 class TestValidateIP(object):
     """IP validation test cases."""
 
-    @pytest.mark.parametrize("ip", ("0.0.0.0", "255.255.255.255", "192.168.1.0"))
+    @pytest.mark.parametrize("ip", ("123.123.123.123", "68.62.43.1", "8.8.8.8"))
     def test_valid(self, ip):
         """Valid ip address values."""
         validate_ip(ip)
 
-    @pytest.mark.parametrize("ip", ("0.0.0.-1", "255.255.255.256", "not an ip address"))
+    @pytest.mark.parametrize("ip", (
+        "0.0.0.-1", "255.255.255.256", "not an ip address"))
     def test_invalid(self, ip):
         """Invalid ip address values."""
         with pytest.raises(ValueError) as exception:
             validate_ip(ip)
         assert str(exception.value) == "Invalid IP address: {!r}".format(ip)
+
+    @pytest.mark.parametrize("ip", ("0.0.0.0", "255.255.255.255", "192.168.1.0"))
+    def test_non_routable(self, ip):
+        """Invalid ip address values."""
+        with pytest.raises(ValueError) as exception:
+            validate_ip(ip)
+        assert str(exception.value) == "Non-Routable IP address: {!r}".format(ip)